refactor : 변수명 수정 및 파일 분리
